refactor(apps)!: move the app exceptions out of core
AppRepositoryError, AppMetadataError and KonfAIAppClientError were defined in
konfai/utils/errors.py and raised there zero times, against 71 raises in
konfai-apps. Core carried its dependent's vocabulary, which is the layering
edge inverted. They now live in konfai_apps/errors.py and still inherit
NamedKonfAIError, which stays core.

konfai_apps.app_repository keeps AppRepositoryError as a module attribute: the
apps contract test asserts that spelling and SlicerKonfAI imports it.

BREAKING CHANGE: the three classes no longer resolve under konfai.utils.errors.
SlicerKonfAI had one stale import of AppRepositoryError from core, beside an
import of the same class from konfai_apps in the same file; the extension must
ship that one-line fix before a konfai without the class reaches its users.
