+++ konfai-apps/konfai_apps/errors.py
@@ -1,0 +1,31 @@
+# Copyright (c) 2025 Valentin Boussot
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+"""Exception types raised by the KonfAI Apps layer."""
+
+from konfai.utils.errors import NamedKonfAIError
+
+
+class AppRepositoryError(NamedKonfAIError):
+    TYPE = "App repository"
+
+
+class AppMetadataError(NamedKonfAIError):
+    TYPE = "Model metadata"
+
+
+class KonfAIAppClientError(NamedKonfAIError):
+    TYPE = "KonfAI App client"
